refactor(sc-agents-common): Revert coreKeynodes file name for resolve include conflicts in apps
