+++ sc-kpm/sc-agents-common/keynodes/coreKeynodes.cpp
@@ -6,7 +6,7 @@
 
 #include <sc-memory/cpp/sc_memory.hpp>
 
-#include "CoreKeynodes.hpp"
+#include "coreKeynodes.hpp"
 
 namespace scAgentsCommon
 {
