+++ sc-kpm/sc-agents-common/keynodes/coreKeynodes.hpp
