+++ sc-kpm/sc-agents-common/sc_agents_common.cpp
@@ -4,7 +4,7 @@
 * (See accompanying file COPYING.MIT or copy at http://opensource.org/licenses/MIT)
 */
 
-#include "keynodes/CoreKeynodes.hpp"
+#include "keynodes/coreKeynodes.hpp"
 #include "agents/ANumberComparison.hpp"
 
 #include "sc_agents_common.hpp"
