+++ sc-kpm/sc-agents-common/utils/AgentUtils.cpp
@@ -11,7 +11,7 @@
 
 #include "AgentUtils.hpp"
 #include "IteratorUtils.hpp"
-#include "keynodes/CoreKeynodes.hpp"
+#include "keynodes/coreKeynodes.hpp"
 
 using namespace std;
 using namespace scAgentsCommon;
