+++ sc-kpm/sc-agents-common/utils/LogicRuleUtils.cpp
@@ -9,7 +9,7 @@
 #include "AgentUtils.hpp"
 #include "IteratorUtils.hpp"
 #include "LogicRuleUtils.hpp"
-#include "keynodes/CoreKeynodes.hpp"
+#include "keynodes/coreKeynodes.hpp"
 
 
 using namespace std;
